+++ changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Upcoming
+## Alpha 4.1.1
 * Wither
   * When he reaches half health and charges, after charging minions will spawn
   * Can no longer heal back past half health, even if kills someone